Crate libcruby_sys

Structs

ANYARGS

ID

VALUE

ruby_value_type
Data types used within the Ruby C interpreter

st_retval
Generally used as return values from traversing callbacks

Statics

Qfalse

Qnil

Qtrue

ST_CHECK
Generally used as a return value from traversing callbacks. It's unclear when this should be used in user-defined code.

ST_CONTINUE
Generally used as a return value from traversing callbacks to indicate that traversing should continue.

ST_DELETE
Generally used as a return value from traversing callbacks to indicate that traversing should stop.

ST_STOP
Generally used as a return value from traversing callbacks to indicate that traversing should stop.

T_ARRAY
This type denotes a Ruby Array.

T_BIGNUM
This type denotes a Ruby Bignum.

T_CLASS
This type denotes a Ruby Class.

T_COMPLEX
This type denotes a Ruby complex number (Complex).

T_DATA
This type denotes a Ruby Data object.

T_FALSE
This type denotes a Ruby false.

T_FILE
This type denotes a Ruby IO.

T_FIXNUM
This type denotes a Ruby Fixnum (31-bit or 63-bit integer).

T_FLOAT
This type denotes a Ruby Float.

T_HASH
This type denotes a Ruby Hash.

T_ICLASS
This type denotes an included Ruby module.

T_IMEMO
This type is generally only for use in Ruby internals

T_MATCH
This type denotes a Ruby MatchData object.

T_MODULE
This type denotes a Ruby Module.

T_NIL
This type denotes a Ruby nil.

T_NODE
This type denotes a Ruby syntax tree node.

T_NONE
This type is generally only for use in Ruby internals

T_OBJECT
This type denotes an ordinary Ruby object.

T_RATIONAL
This type denotes a Ruby rational number (Rational).

T_REGEXP
This type denotes a Ruby regular expression (Regexp).

T_STRING
This type denotes a Ruby String.

T_STRUCT
This type denotes a Ruby Struct.

T_SYMBOL
This type denotes a Ruby Symbol.

T_TRUE
This type denotes a Ruby true.

T_UNDEF
This type denotes undefined.

T_ZOMBIE
This type denotes a Ruby object awaiting finalization.

rb_cArray
The Array class

rb_cBasicObject
The BasicObject class

rb_cBinding
The Binding class

rb_cClass
The Class class

rb_cComplex
The Complex class

rb_cDir
The Dir class

rb_cEncoding
The Encoding class

rb_cEnumerator
The Enumerator class

rb_cFalseClass
The FalseClass class

rb_cFile
The File class

rb_cFloat
The Float class

rb_cHash
The Hash class

rb_cIO
The IO class

rb_cInteger
The Integer class

rb_cMatch
The MatchData class

rb_cMethod
The Method class

rb_cModule
The Module class

rb_cNilClass
The NilClass class

rb_cNumeric
The Numeric class

rb_cObject
The Object class

rb_cProc
The Proc class

rb_cRandom
The Random class

rb_cRange
The Range class

rb_cRational
The Rational class

rb_cRegexp
The Regexp class

rb_cStat
The File::Stat class

rb_cString
The String class

rb_cStruct
The Struct class

rb_cSymbol
The Symbol class

rb_cThread
The Thread class

rb_cTime
The Time class

rb_cTrueClass
The TrueClass class

rb_cUnboundMethod
The UnboundMethod class

rb_eArgError
The ArgumentError exception class

rb_eEOFError
The EOFError exception class

rb_eEncCompatError
The Encoding::CompatibilityError exception class

rb_eEncodingError
The EncodingError exception class

rb_eException
The Exception class

rb_eFloatDomainError
The FloatDomainError exception class

rb_eIOError
The IOError exception class

rb_eIndexError
The IndexError exception class

rb_eInterrupt
The Interrupt exception class

rb_eKeyError
The KeyError exception class

rb_eLoadError
The LoadError exception class

rb_eLocalJumpError
The LocalJumpError exception class

rb_eMathDomainError
The Math::DomainError exception class

rb_eNameError
The NameError exception class

rb_eNoMemError
The NoMemoryError exception class

rb_eNoMethodError
The NoMethodError exception class

rb_eNotImpError
The NotImplementedError exception class

rb_eRangeError
The RangeError exception class

rb_eRegexpError
The RegexpError exception class

rb_eRuntimeError
The RuntimeError exception class

rb_eScriptError
The ScriptError exception class

rb_eSecurityError
The SecurityError exception class

rb_eSignal
The Signal exception class

rb_eStandardError
The StandardError exception class

rb_eStopIteration
The StopIteration exception class

rb_eSyntaxError
The SyntaxError exception class

rb_eSysStackError
The SystemStackError exception class

rb_eSystemCallError
The SystemCallError exception class

rb_eSystemExit
The SystemExit exception class

rb_eThreadError
The ThreadError exception class

rb_eTypeError
The TypeError exception class

rb_eZeroDivError
The ZeroDivisionError exception class

rb_mComparable
The Comparable module

rb_mEnumerable
The Enumerable module

rb_mErrno
The Errno module

rb_mFileTest
The FileTest module

rb_mGC
The GC module

rb_mKernel
The Kernel module

rb_mMath
The Math module

rb_mProcess
The Process module

rb_mWaitReadable
The IO::WaitReadable module

rb_mWaitWritable
The IO::WaitWritable module

Functions

CLASS_OF⚠
Returns the Ruby class of the object.

DBL2NUM⚠
Converts a double to a Ruby Float.

INT2NUM⚠
Converts an integer to a Ruby Numeric.

LL2NUM⚠
Converts a long long to a Ruby Numeric.

LONG2NUM⚠
Converts a long to a Ruby Numeric.

NUM2DBL⚠
Converts a Ruby Numeric to an unsigned long long.

NUM2INT⚠
Converts a Ruby Numeric to an integer.

NUM2LL⚠
Converts a Ruby Numeric to a long long.

NUM2LONG⚠
Converts a Ruby Numeric to a long.

NUM2UINT⚠
Converts a Ruby Numeric to an unsigned integer.

NUM2ULL⚠
Converts a Ruby Numeric to an unsigned long long.

NUM2ULONG⚠
Converts a Ruby Numeric to an unsigned long.

RARRAY_LEN⚠
Returns the number of elements in the Ruby Array.

RB_TYPE_P⚠
Returns a C boolean (zero if false, non-zero if true) indicating whether the object is of the internal Ruby type.

RHASH_SIZE⚠
Returns the number of elements in the Ruby Hash.

RSTRING_LEN⚠
Returns the byte length of the Ruby String.

RSTRING_PTR⚠
Returns a pointer to the Ruby String data.

UINT2NUM⚠
Converts an unsigned integer to a Ruby Numeric.

ULL2NUM⚠
Converts an unsigned long long to a Ruby Numeric.

ULONG2NUM⚠
Converts an unsigned long to a Ruby Numeric.

rb_ary_entry⚠
Returns the element at the given index or nil if the index is out-of-bounds for the array.

rb_ary_new⚠
Constructs a new, empty array.

rb_ary_new_capa⚠
Constructs a new, empty array with the specified capacity.

rb_ary_push⚠
Pushes an item on to the end of an array, returning the array itself.

rb_block_given_p⚠
Returns a C boolean (zero if false, non-zero if true) indicated whether yield would execute a block in the current method.

rb_class_new_instance⚠
Constructs a new instance of a class by calling its allocator and constructor (alloc and initialize) as ::new normally would.

rb_const_get⚠
Fetches a constant from a module or class.

rb_define_class⚠
Defines a new class.

rb_define_method⚠
Defines a public method on a class.

rb_define_module_function⚠
Defines a method on a module, both on the module itself and as a private method for use by anything including the module.

rb_define_singleton_method⚠
Defines a singleton method on a class.

rb_enc_get_index⚠
Returns the encoding index of the provided Ruby object

rb_hash_aset⚠
Inserts a key-value pair into the hash.

rb_hash_foreach⚠
Executes a function on each key-value pair in a hash.

rb_hash_new⚠
Constructs a new, empty hash.

rb_id2str⚠
Convert an ID to a Ruby String.

rb_id2sym⚠
Convert an ID to a Symbol.

rb_inspect⚠
Returns a human-readable Ruby string representation of an object, similarly to Ruby's Object#inspect.

rb_intern⚠
Converts an ASCII-encoded, nul-terminated C string to an ID.

rb_intern2⚠
Converts an ASCII-encoded C string of the given length to an ID.

rb_intern_str⚠
Convert a string to an ID.

rb_obj_class⚠
Looks up the nearest ancestor class of the object, skipping singleton classes or module inclusions.

rb_obj_classname⚠
Gets the object's class' name

rb_sym2id⚠
Convert a Symbol to an ID.

rb_undef_method⚠
Undefines the named method on a class.

rb_utf8_encindex⚠
Returns the encoding index for UTF-8

rb_utf8_str_new⚠
Constructs a new Ruby string from a UTF-8 encoded C string of a given length.