Function libcruby_sys::rb_define_module_function

pub unsafe extern "C" fn rb_define_module_function(
    module: VALUE, 
    name: *const c_char, 
    func: ANYARGS<VALUE>, 
    arity: c_int
)

Defines a method on a module, both on the module itself and as a private method for use by anything including the module.

module - a Module

See rb_define_method for additional details on arguments.

Safety

See rb_define_method.

Ruby Documentaiton

2.5

Defined In

2.3: ruby.h | class.c

2.4: ruby.h | class.c

2.5: ruby.h | class.c

2.6: ruby.h | class.c