Function libcruby_sys::rb_inspect

pub unsafe extern "C" fn rb_inspect(obj: VALUE) -> VALUE

Returns a human-readable Ruby string representation of an object, similarly to Ruby's Object#inspect.

Unlike Object#inspect, it escapes characters to keep the result compatible to the default internal or external encoding. If the default internal or external encoding is ASCII compatible, the encoding of the inspected result must be compatible with it. If the default internal or external encoding is ASCII incompatible, the result must be ASCII only.

obj - any Ruby object

Returns a String

Safety

Exceptions

May call user-defined code that could raise an exception

Defined In

2.3: intern.h

2.4: intern.h

2.5: intern.h | object.c

2.6: intern.h